Make changes following the GithubAdvices #JC19-258_
