+++ vanila/package.json
@@ -10,6 +10,7 @@
   },
   "devDependencies": {
     "@chernodub/eslint-config": "^2.0.2",
+    "@types/node": "^17.0.13",
     "@typescript-eslint/eslint-plugin": "^5.10.1",
     "@typescript-eslint/parser": "^5.10.1",
     "eslint": "^8.7.0",
@@ -17,12 +18,13 @@
     "eslint-plugin-jsdoc": "^37.7.0",
     "eslint-plugin-prefer-arrow": "^1.2.3",
     "prettier": "^2.5.1",
-    "typescript": "^4.5.5",
-    "vite": "^2.7.2",
+    "stylelint": "^14.3.0",
     "stylelint-config-recommended": "^6.0.0",
-    "stylelint": "^14.3.0"
+    "typescript": "^4.5.5",
+    "vite": "^2.7.2"
   },
   "dependencies": {
+    "dotenv": "^15.0.0",
     "firebase": "^9.4.1"
   }
 }
