+++ vanila/src/DTO/collectionDTO.ts
@@ -1,0 +1,20 @@
+/**
+ * Parent interface for db collections.
+ */
+export interface CollectionDTO<T> {
+
+  /**
+   * Id of item.
+   */
+  readonly pk: number;
+
+  /**
+   *Model of item.
+   */
+  readonly model: string;
+
+  /**
+   * Fields of collection.
+   */
+  readonly fields: T;
+}
