+++ vanila/src/DTO/peopleDTO.ts
@@ -1,0 +1,60 @@
+/**
+ * Characteristics of people.
+ */
+export interface PeopleDTO {
+
+  /**
+   * Birthday year.
+   */
+  birthYear: string;
+
+  /**
+   *Date of creation.
+   */
+  created: string;
+
+  /**
+   * Date of edition.
+   */
+  edited: string;
+
+  /**
+   * Color of eyes.
+   */
+  eyeColor: string;
+
+  /**
+   * Gender.
+   */
+  gender: string;
+
+  /**
+   * Color of Hair.
+   */
+  hairColor: string;
+
+  /**
+   * Height in sm.
+   */
+  height: string;
+
+  /**
+   * Primary key of homeland planet.
+   */
+  homeWorld: number;
+
+  /**
+   * Mass of people.
+   */
+  mass: string;
+
+  /**
+   * Name op people.
+   */
+  name: string;
+
+  /**
+   * Color of skin.
+   */
+  skinColor: string;
+}
