+++ vanila/src/DTO/speciesDTO.ts
@@ -1,0 +1,70 @@
+/**
+ * Planet characteristics.
+ */
+export interface SpeciesDTO {
+
+  /**
+   * Species average height.
+   */
+  averageHeight: string;
+
+  /**
+   * Average lifespan of species.
+   */
+  averageLifespan: string;
+
+  /**
+   * Classification of species.
+   */
+  classification: string;
+
+  /**
+   * Date of creation.
+   */
+  created: string;
+
+  /**
+   *Species designation.
+   */
+  designation: string;
+
+  /**
+   * Date of edition.
+   */
+  edited: string;
+
+  /**
+   * Eye colors.
+   */
+  eyeColors: string;
+
+  /**
+   * Hair color.
+   */
+  hairColors: string;
+
+  /**
+   * Primary key of homeland planet.
+   */
+  homeWorld: number;
+
+  /**
+   * Language of species.
+   */
+  language: string;
+
+  /**
+   * Name of species.
+   */
+  name: string;
+
+  /**
+   * Array of primary keys of people.
+   */
+  people: number[];
+
+  /**
+   * Skin color.
+   */
+  skinColors: string;
+}
