+++ vanila/src/DTO/starshipDTO.ts
@@ -1,0 +1,20 @@
+/**
+ * Starships characteristics.
+ */
+export interface StarshipDTO {
+
+  /**
+   *MLGT.
+   */
+  MLGT: string;
+
+  /**
+   * Hyperdrive rating.
+   */
+  hyperdriveRating: string;
+
+  /**
+   * Class of starship.
+   */
+  starhipClass: string;
+}
