+++ vanila/src/DTO/vehicleDTO.ts
@@ -1,0 +1,10 @@
+/**
+ * Vehicle characteristics.
+ */
+export interface VehicleDTO {
+
+  /**
+   * Class of vehicle.
+   */
+  vehicleClass: string;
+}
