--- vanila/src/interfaces/collectionInterface.ts
@@ -1,15 +1,0 @@
-/**
- * Parent interface for db collections.
- */
-export interface CollectionInterface {
-
-  /**
-   * Id of item.
-   */
-  pk: number;
-
-  /**
-   *Model of item.
-   */
-  model: string;
-}
