--- vanila/src/interfaces/peopleInterface.ts
@@ -1,68 +1,0 @@
-import { CollectionInterface } from './collectionInterface';
-
-/**
- * Characteristics of people.
- */
-export interface PeopleInterface extends CollectionInterface {
-
-  /**
-   * Fields of people characteristics.
-   */
-  fields: {
-
-    /**
-     * Birthday year.
-     */
-    birthYear: string;
-
-    /**
-     *Date of creation.
-     */
-    created: string;
-
-    /**
-     * Date of edition.
-     */
-    edited: string;
-
-    /**
-     * Color of eyes.
-     */
-    eyeColor: string;
-
-    /**
-     * Gender.
-     */
-    gender: string;
-
-    /**
-     * Color of Hair.
-     */
-    hairColor: string;
-
-    /**
-     * Height in sm.
-     */
-    height: string;
-
-    /**
-     * Primary key of homeland planet.
-     */
-    homeWorld: number;
-
-    /**
-     * Mass of people.
-     */
-    mass: string;
-
-    /**
-     * Name op people.
-     */
-    name: string;
-
-    /**
-     * Color of skin.
-     */
-    skinColor: string;
-  };
-}
