--- vanila/src/interfaces/starshipInterface.ts
@@ -1,28 +1,0 @@
-import { CollectionInterface } from './collectionInterface';
-
-/**
- * Starships characteristics.
- */
-export interface StarshipInterface extends CollectionInterface{
-
-  /**
-   * Fields of characteristics.
-   */
-  fields: {
-
-    /**
-     *MLGT.
-     */
-    MLGT: string;
-
-    /**
-     * Hyperdrive rating.
-     */
-    hyperdriveRating: string;
-
-    /**
-     * Class of starship.
-     */
-    starhipClass: string;
-  };
-}
