--- vanila/src/interfaces/transportInterface.ts
@@ -1,73 +1,0 @@
-import { CollectionInterface } from './collectionInterface';
-
-/**
- * Transport characteristics.
- */
-export interface TransportInterface extends CollectionInterface {
-
-  /**
-   * Transport characteristic fields.
-   */
-  fields: {
-
-    /**
-     * Cargo capacity.
-     */
-    cargoCapacity: string;
-
-    /**
-     * Consumables.
-     */
-    consumables: string;
-
-    /**
-     * Price in credits.
-     */
-    costInCredits: string;
-
-    /**
-     * Date of creation.
-     */
-    created: string;
-
-    /**
-     *Crew id.
-     */
-    crew: string;
-
-    /**
-     * Edited date.
-     */
-    edited: string;
-
-    /**
-     * Length of transport.
-     */
-    length: string;
-
-    /**
-     * Name of manufacturer.
-     */
-    manufacturer: string;
-
-    /**
-     * Max atmospheric speed.
-     */
-    maxAtmospheringSpeed: string;
-
-    /**
-     * Transport model.
-     */
-    model: string;
-
-    /**
-     * Transport name.
-     */
-    name: string;
-
-    /**
-     *Quality of passengers.
-     */
-    passengers: string;
-  };
-}
