--- vanila/src/interfaces/vehicleInterface.ts
@@ -1,18 +1,0 @@
-import { CollectionInterface } from './collectionInterface';
-
-/**
- * Vehicle characteristics.
- */
-export interface VehicleInterface extends CollectionInterface{
-
-  /**
-   * Characteristic fieds.
-   */
-  fields: {
-
-    /**
-     * Class of vehicle.
-     */
-    vehicleClass: string;
-  };
-}
